moved low level qa functions to a separate library file
